more bugfixes and merge unit tests and int tests
